Split generator tests into domain specific files
